pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-tts>=0.3.1",
    "docx>=0.2.4",
    "google-generativeai>=0.8.4",
    "gtts>=2.5.4",
    "numpy==1.24.3",
    "openai>=1.61.1",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "pyttsx3>=2.98",
    "scipy>=1.15.1",
    "soundfile>=0.13.1",
    "streamlit>=1.42.0",
    "torch>=2.6.0",
    "trafilatura>=2.0.0",
    "transformers>=4.48.3",
    "tts==0.17.2",
]

[[tool.uv.index]]
explicit = true
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
